AOP公共字段填充

+++ sky-server/src/main/java/com/sky/service/impl/CategoryServiceImpl.java
@@ -3,17 +3,19 @@ package com.sky.service.impl;
 import com.github.pagehelper.Page;
 import com.github.pagehelper.PageHelper;
 import com.sky.constant.StatusConstant;
-import com.sky.context.BaseContext;
 import com.sky.dto.CategoryDTO;
 import com.sky.dto.CategoryPageQueryDTO;
 import com.sky.entity.Category;
+import com.sky.entity.Dish;
+import com.sky.entity.Setmeal;
 import com.sky.mapper.CategoryMapper;
+import com.sky.mapper.DishMapper;
+import com.sky.mapper.SetmealMapper;
 import com.sky.result.PageResult;
 import com.sky.service.CategoryService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.time.LocalDateTime;
 import java.util.List;
 
 
@@ -22,11 +24,15 @@ public class CategoryServiceImpl implements CategoryService {
 
     @Autowired
     private CategoryMapper categoryMapper;
+    @Autowired
+    private DishMapper dishMapper;
+    @Autowired
+    private SetmealMapper setmealMapper;
 
     /**
      * 分页查询分类信息
-     * @param categoryPageQueryDTO
-     * @return
+     * @param categoryPageQueryDTO 分页查询参数
+     * @return 返回分页结果
      */
     public PageResult SelectByPage(CategoryPageQueryDTO categoryPageQueryDTO) {
         PageHelper.startPage(categoryPageQueryDTO.getPage(), categoryPageQueryDTO.getPageSize());
@@ -34,32 +40,25 @@ public class CategoryServiceImpl implements CategoryService {
         return new PageResult(page.getTotal(), page.getResult());
     }
 
-    /**
+     /**
      * 新增分类
-     * @param categoryDTO
+     * @param categoryDTO 分类信息
      */
     public void addCategory(CategoryDTO categoryDTO) {
-        LocalDateTime now = LocalDateTime.now();
-        Long idConstant = BaseContext.getCurrentId();
         Category category =Category.builder()
                 .id(categoryDTO.getId())
                 .type(categoryDTO.getType())
                 .name(categoryDTO.getName())
                 .sort(categoryDTO.getSort())
-                .createTime(now)
-                .updateTime(now)
-                .updateUser(idConstant)
-                .createUser(idConstant)
                 .status(StatusConstant.ENABLE)
                 .build();
         categoryMapper.addCategory(category);
     }
 
-
     /**
      * 根据类型查询分类信息
-     * @param type
-     * @return
+     * @param type 分类类型
+     * @return 返回分类列表
      */
     public List<Category> SelectByType(Integer type) {
         List<Category> categories = categoryMapper.SelectByType(type);
@@ -68,43 +67,41 @@ public class CategoryServiceImpl implements CategoryService {
 
     /**
      * 根据id删除分类
-     * @param id
+     * @param id 分类id
      */
     public void DeleteById(Long id) {
+        //分类下存在菜品或者套餐则不能删除
+        Dish dishCount = dishMapper.QueryById(id);
+        Setmeal setmealCount = setmealMapper.QueryById(id);
+        if(dishCount != null || setmealCount != null){
+            throw new IllegalArgumentException("分类下存在菜品或者套餐，不能删除");
+        }
         categoryMapper.DeleteById(id);
     }
 
-    /**
+     /**
      * 启用或停用分类
-     * @param status
-     * @param id
+     * @param status 状态
+     * @param id 分类id
      */
     public void OnOrStop(Integer status, Long id) {
-        LocalDateTime now = LocalDateTime.now();
-        Long idConstant = BaseContext.getCurrentId();
         Category category = Category.builder()
                 .id(id)
                 .status(status)
-                .updateTime(now)
-                .updateUser(idConstant) // TODO 从线程中获取用户id常量
                 .build();
         categoryMapper.Update(category);
     }
 
-    /**
+     /**
      * 更新分类
-     * @param categoryDTO
+     * @param categoryDTO 分类信息
      */
     public void Update(CategoryDTO categoryDTO) {
-        LocalDateTime now = LocalDateTime.now();
-        Long idConstant = BaseContext.getCurrentId();
         Category category =Category.builder()
                 .id(categoryDTO.getId())
                 .type(categoryDTO.getType())
                 .name(categoryDTO.getName())
                 .sort(categoryDTO.getSort())
-                .updateTime(now)
-                .updateUser(idConstant) // TODO 从线程中获取用户id常量
                 .build();
         categoryMapper.Update(category);
     }
